Refactored api connector file for articles
