+++ app/javascript/src/hooks/reactQuery/analytics/useAnalytics.js
@@ -1,7 +1,7 @@
 import { prop } from "ramda";
 import { useQuery, useMutation } from "react-query";
 
-import articlesApi from "apis/articles";
+import analyticsApi from "apis/articles/analytics";
 import queryClient from "utils/queryClient";
 
 const onMutation = () => queryClient.invalidateQueries(["generatePdf"]);
@@ -9,7 +9,7 @@ const onMutation = () => queryClient.invalidateQueries(["generatePdf"]);
 export const useFetchArticleAnalytics = ({ currentPageNumber }) =>
   useQuery(
     ["analytics", currentPageNumber],
-    async () => await articlesApi.analytics(currentPageNumber),
+    async () => await analyticsApi.fetch(currentPageNumber),
     {
       select: prop("data"),
       onError: error => logger.log(error),
@@ -17,7 +17,7 @@ export const useFetchArticleAnalytics = ({ currentPageNumber }) =>
   );
 
 export const useGeneratePdf = () =>
-  useMutation(["generatePdf"], async () => await articlesApi.generatePdf(), {
+  useMutation(["generatePdf"], async () => await analyticsApi.generatePdf(), {
     onSuccess: onMutation,
     onError: error => logger.log(error),
   });
