+++ app/javascript/src/hooks/reactQuery/articles/actions/edit/useRestoreVersion.js
@@ -1,13 +1,13 @@
 import { useMutation } from "react-query";
 
-import articlesApi from "apis/articles";
+import versionsApi from "apis/articles/versions";
 import queryClient from "utils/queryClient";
 
 const onMutation = () =>
   queryClient.invalidateQueries(["dashboard.article", "dashboard.categories"]);
 
 const handleRestore = async ({ article, version }) =>
-  await articlesApi.restore({ id: article.id, versionId: version.id });
+  await versionsApi.restore({ id: article.id, versionId: version.id });
 
 export const useRestoreVersion = ({ refetch, setShowVersionHistory }) =>
   useMutation(handleRestore, {
